+++ packages/fabric-cqrs/src/store/query/__tests__/query.store.unit-test.ts
@@ -459,7 +459,6 @@ describe('Store/query Test', () => {
           _ts: 1590739000,
           _creator: 'org1-admin',
           _created: 1590738792,
-          _organization: null,
         },
         {
           value: 3,
@@ -469,7 +468,6 @@ describe('Store/query Test', () => {
           _ts: 1590740002,
           _creator: 'org1-admin',
           _created: 1590740000,
-          _organization: null,
         },
         {
           value: 2,
@@ -479,7 +477,6 @@ describe('Store/query Test', () => {
           _ts: 1590740004,
           _creator: 'org1-admin',
           _created: 1590740003,
-          _organization: null,
         },
       ]);
     }));
@@ -564,7 +561,6 @@ describe('Store/query Test', () => {
           _ts: 1590739000,
           _created: 1590738792,
           _creator: 'org1-admin',
-          _organization: null,
         },
         {
           value: 3,
@@ -574,7 +570,6 @@ describe('Store/query Test', () => {
           _ts: 1590740002,
           _created: 1590740000,
           _creator: 'org1-admin',
-          _organization: null,
         },
       ]);
     }));