+++ packages/fabric-cqrs/src/types/commit.ts
@@ -35,6 +35,56 @@ export type BaseEvent = {
 };
 
 /**
+ * System event name - creating private data
+ * @ignore
+ */
+export const TRACK_EVENT = 'PrivateDataTracked';
+
+/**
+ * System field name - private data org list
+ * @ignore
+ */
+export const TRACK_FIELD = '_privateData';
+
+export const TRACK_FIELD_S = 'privateData';
+
+/**
+ * System field name - org list
+ * @ignore
+ */
+export const ORGAN_FIELD = '_organization';
+
+// export const ORGAN_FIELD_S = 'organization';
+
+/**
+ * System field name - commit timestamp
+ * @ignore
+ */
+export const TS_FIELD = '_ts';
+
+// export const TS_FIELD_C = 'ts';
+
+// export const TS_FIELD_O = 'modifiedAt';
+
+/**
+ * System field name - enrollment ID used when creating the entity
+ * @ignore
+ */
+export const CREATOR_FIELD = '_creator';
+
+// export const CREATOR_FIELD_S = 'creator';
+
+/**
+ * System field name - timestamp when creating the entity
+ * @ignore
+ */
+export const CREATED_FIELD = '_created';
+
+// export const CREATED_FIELD_C = 'created';
+
+// export const CREATED_FIELD_O = 'createdAt';
+
+/**
  * @about Commit
  */
 export type Commit = {