+++ packages/gateway-lib/src/common/model/organization/domain/events.ts
@@ -5,9 +5,9 @@ export interface OrgStarted extends BaseEvent {
   payload: {
     mspId: string;
     timestamp: number;
-    _ts?: number;
-    _created?: number;
-    _creator?: string;
+    // _ts?: number;
+    // _created?: number;
+    // _creator?: string;
   };
 }
 
@@ -17,9 +17,9 @@ export interface OrgNameDefined extends BaseEvent {
     mspId: string;
     name: string;
     timestamp: number;
-    _ts?: number;
-    _created?: number;
-    _creator?: string;
+    // _ts?: number;
+    // _created?: number;
+    // _creator?: string;
   };
 }
 
@@ -29,9 +29,9 @@ export interface OrgUrlDefined extends BaseEvent {
     mspId: string;
     url: string;
     timestamp: number;
-    _ts?: number;
-    _created?: number;
-    _creator?: string;
+    // _ts?: number;
+    // _created?: number;
+    // _creator?: string;
   };
 }
 
@@ -40,9 +40,9 @@ export interface OrgDowned extends BaseEvent {
   payload: {
     mspId: string;
     timestamp: number;
-    _ts?: number;
-    _created?: number;
-    _creator?: string;
+    // _ts?: number;
+    // _created?: number;
+    // _creator?: string;
   };
 }