+++ packages/gateway-lib/src/common/model/organization/service/resolvers.ts
@@ -1,6 +1,5 @@
 import util from 'util';
-import { TRACK_FIELD, ORGAN_NAME } from '@fabric-es/fabric-cqrs';
-import { Organization, OrgContext } from '..';
+import { ORGAN_NAME, OrgContext } from '..';
 import { catchResolverErrors } from '../../../..';
 import { getLogger } from '../../../../utils';