+++ packages/gateway-lib/src/types/index.ts
@@ -8,3 +8,4 @@ export * from './notification';
 export * from './dataSrc';
 export * from './commandHandler';
 export * from './federatedService';
+export * from './auth0UserInfo';