+++ packages/gateway-lib/src/utils/index.ts
@@ -8,3 +8,5 @@ export * from './composeRedisRepos';
 export * from './waitForSecond';
 export * from './catchResolverErrors';
 export * from './createRemoteData';
+export * from './buildRedisOptions';
+export * from './createGatewayWithAuth0';